file encryption modularized
